@@ -9,7 +9,7 @@ buildscript {
 }
 
 plugins {
-    id 'com.onesignal.androidsdk.onesignal-gradle-plugin' version '0.8.1'
+    id 'com.onesignal.androidsdk.onesignal-gradle-plugin' version '0.12.0'
 }
 
 apply plugin: 'com.android.application'
@@ -24,7 +24,7 @@ repositories {
 
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "mandela.cct.ansteph.kazihealth"
 
@@ -35,11 +35,11 @@ android {
         ]
 
         minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 5
-        versionName "1.5"
+        targetSdkVersion 29
+        versionCode 7
+        versionName "1.7"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -58,10 +58,10 @@ android {
             dimension "cct"
             minSdkVersion 19
             applicationId 'mandela.cct.ansteph.kazihealth'
-            targetSdkVersion 28
-            testInstrumentationRunner 'android.support.test.runner.AndroidJUnitRunner'
-            versionCode 5
-            versionName '1.5'
+            targetSdkVersion 29
+            testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+            versionCode 7
+            versionName '1.7'
 
         }
     }
@@ -69,33 +69,33 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.0'
-    implementation 'com.android.support:design:28.0.0'
-    compile 'de.hdodenhof:circleimageview:2.1.0'
-    compile 'com.yarolegovich:lovely-dialog:1.1.0'
-    implementation 'com.google.firebase:firebase-core:11.4.2'
-    implementation 'com.google.firebase:firebase-database:11.4.2'
-    implementation 'com.google.firebase:firebase-messaging:11.4.2'
-    compile 'com.google.firebase:firebase-auth:11.4.2'
-    compile 'com.google.android.gms:play-services-auth:11.4.2'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
-    compile 'com.milaptank:stv:1.0.0'
-    compile project(':android-pdf-viewer')
+    implementation 'de.hdodenhof:circleimageview:2.2.0'
+    implementation 'com.yarolegovich:lovely-dialog:1.1.0'
+    implementation 'com.google.firebase:firebase-core:17.2.0'
+    implementation 'com.google.firebase:firebase-database:19.1.0'
+    implementation 'com.google.firebase:firebase-messaging:20.0.0'
+    implementation 'com.google.firebase:firebase-auth:19.1.0'
+    implementation 'com.google.android.gms:play-services-auth:17.0.0'
 
-    implementation 'com.onesignal:OneSignal:[3.9.1, 3.99.99]'
+    implementation 'com.milaptank:stv:1.0.0'
+    implementation project(':android-pdf-viewer')
+
+    implementation 'com.onesignal:OneSignal:[3.11.4, 3.99.99]'
 
     implementation('com.shagi:material-datepicker:1.3') {
         exclude group: 'com.android.support'
     }
-    implementation 'com.android.volley:volley:1.0.0'
+    implementation 'com.android.volley:volley:1.1.0'
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
-    compile('com.crashlytics.sdk.android:crashlytics:2.9.2@aar') {
-        transitive = true;
+    androidTestImplementation 'androidx.test:runner:1.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
+    implementation('com.crashlytics.sdk.android:crashlytics:2.9.2@aar') {
+        transitive = true
     }
 }
 apply plugin: 'com.google.gms.google-services'
