@@ -2,11 +2,9 @@ package mandela.cct.ansteph.kazihealth.view.firebasereg;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Patterns;
 import android.view.View;
 import android.widget.EditText;
@@ -21,7 +19,6 @@ import com.google.firebase.auth.FirebaseAuth;
 import mandela.cct.ansteph.kazihealth.R;
 import mandela.cct.ansteph.kazihealth.app.GlobalRetainer;
 import mandela.cct.ansteph.kazihealth.view.profile.RiskProfile;
-import mandela.cct.ansteph.kazihealth.view.register.Login;
 
 public class Login_Firebase extends AppCompatActivity {
 
