@@ -3,11 +3,10 @@ package mandela.cct.ansteph.kazihealth.view.register;
 import android.app.DatePickerDialog;
 import android.app.Dialog;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.DialogFragment;
-import android.widget.AutoCompleteTextView;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
+
 import android.widget.DatePicker;
-import android.widget.EditText;
 import android.widget.TextView;
 
 import java.util.Calendar;
